requirements.txt
openai
tiktoken
python-dotenv
python-telegram-bot
google-cloud-speech
